starting server in run.sh

+++ run.sh
@@ -11,11 +11,13 @@ fi
 
 ConsumerApp="build/consumer/ConsumerApp"
 PruducerApp="build/producer/ProducerApp"
+ServerApp="build/server/ServerApp"
 
-if [[ -f $ConsumerApp && -f $PruducerApp ]]
+if [[ -f $ConsumerApp && -f $PruducerApp && -f $ServerApp ]]
 then
     cp $ConsumerApp $FOLDER
     cp $PruducerApp $FOLDER
+    cp $ServerApp $FOLDER
 else
     echo "Build the project first!"
     exit 1
@@ -25,5 +27,6 @@ fi
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 
+x-terminal-emulator -e "${SCRIPT_DIR}/${FOLDER}/ServerApp" ServerApp &
 x-terminal-emulator -e "${SCRIPT_DIR}/${FOLDER}/ConsumerApp" ConsumerApp &
 x-terminal-emulator -e "${SCRIPT_DIR}/${FOLDER}/ProducerApp" ProducerApp &
